+++ ext_localconf.php
@@ -4,16 +4,9 @@ if (!defined('TYPO3_MODE')) {
     die('Access denied.');
 }
 
-$versionParts = explode('.', TYPO3_version);
-$rnbaseExtPath = (intval($versionParts[0]) >= 6) ?
-    \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::extPath('rn_base') :
-    t3lib_extMgm::extPath('rn_base');
-require_once $rnbaseExtPath.'class.tx_rnbase.php';
-tx_rnbase::load('tx_rnbase_util_Typo3Classes');
 
 if (!is_array($GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations']['rnbase']) &&
     Tx_Rnbase_Configuration_Processor::getExtensionCfgValue('rn_base', 'activateCache')) {
-    tx_rnbase::load('tx_rnbase_cache_Manager');
     tx_rnbase_cache_Manager::registerCache(
         'rnbase',
         tx_rnbase_cache_Manager::CACHE_FRONTEND_VARIABLE,